backend/requirements.txt
fastapi
uvicorn
python-multipart
pandas
python-dotenv
langchain
langgraph
langchain-openai
langchain-text-splitters
langchain-core
sqlalchemy
chromadb
openpyxl
python-docx
PyPDF2
sentence-transformers
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pytest
pytest-asyncio
httpx
openai